Alteração da localização do botão sair e adição do botão voltar
