+++ src/components/Header/index.jsx
@@ -1,22 +1,20 @@
 import { useNavigate } from "react-router-dom";
-import { removeToken } from "@/utils/api";
 
 const Header = (props) => {
-  const { title, userTitle } = props;
+  const { title } = props;
 
   const navigate = useNavigate();
 
-  const logout = () => {
-    removeToken();
-    navigate("/login");
+  const voltar = () => {
+      window.history.back();
   };
 
   return (
     <div className="header d-flex align-items-center justify-content-between my-4 col-lg-12">
       <h2>{title}</h2>
       <div className="d-flex align-items-center">
-        <span className="font-size-20">{userTitle}</span>
-        <button className="btn btn-secondary ms-3" onClick={logout}>
+        <span className="font-size-20">Voltar</span>
+        <button className="btn btn-secondary ms-3" onClick={voltar}>
           <i className="fa fa-right-from-bracket" />
         </button>
       </div>
